Added permissions to Fossil changes, fixed bug with image upload.
Previously, if an image was uploaded that exceeded the maximum filesize
set by the server environment, the error resulting would erroneously alert
the user that they need to upload an image of a certain filetype, even if that
filetype was provided.

The new error message indicates what the issue may be.

Closes #56.