BUG: Taxon helper no longer overwrites common name
Now the common name is not overwriten to null when the taxonomy helper popup is
used.

I think that there might should be an alert to say that the common name may no
longer be accurate once this is changed. If I have time, I'll try to add this
to the pull request.

Closes myfossil/myfossil-specimen#81